// src/interfaces/VCtasctes.tsx
export interface VCtasctes {
    deudaOriginal: number;
    intereses: number;
    nro_cedulon_paypertic: number;
    pago_parcial: boolean;
    pago_a_cuenta: number;
    nro_transaccion: number;
    periodo: string;
    importe: number;
    fecha_vencimiento: string;
    categoria_deuda: number;
  }